fix reported memory var overflow by using unsigned long long
git-svn-id: https://conky.svn.sourceforge.net/svnroot/conky/trunk/conky1@775 7f574dfc-610e-0410-a909-a81674777703

+++ src/conky.h
@@ -242,8 +242,8 @@ struct information {
 	double uptime;
 
 	/* memory information in kilobytes */
-	unsigned long mem, memmax, swap, swapmax;
-	unsigned long bufmem, buffers, cached;
+	unsigned long long mem, memmax, swap, swapmax;
+	unsigned long long bufmem, buffers, cached;
 
 	unsigned short procs;
 	unsigned short run_procs;
